Fix clippy warning + regen highlight keywords
- `crates/harn-vm/src/stdlib/postgres.rs`: drop `as i64` cast on
  `VmValue::Duration(ms)` — `ms` is already i64, clippy::unnecessary_cast.
- `docs/theme/harn-keywords.js`: regenerated via `make gen-highlight`
  to include the new HTTP server, multipart, cookie/session, signed_url,
  WebSocket server, SSE server, and Postgres builtins from the codex
  stack (#649–#656).

+++ crates/harn-vm/src/stdlib/postgres.rs
@@ -399,7 +399,7 @@ fn bind_params<'q>(
             VmValue::Float(value) => query.bind(*value),
             VmValue::String(value) => query.bind(value.to_string()),
             VmValue::Bytes(value) => query.bind((**value).clone()),
-            VmValue::Duration(ms) => query.bind(*ms as i64),
+            VmValue::Duration(ms) => query.bind(*ms),
             value => query.bind(sqlx_core::types::Json(vm_value_to_json(value))),
         };
     }
